backurne: limit the creation of snaps
Snapshots can be a burden on the cluster, if they are left proliferate.
When an image reaches hundreds of snapshots, and those are bulk-removed,
OSDs will take of lot of CPU to walk through.
CPU load as well as rocksdb pressure may have impact on the clients.
We try to avoid this behavior by denying the creation of a snapshot,
if there is already too much for our usecase.

Signed-off-by: Alexandre Bruyelles <jack@jack.fr.eu.org>